requirements/main.txt
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
pycparser==2.20
python-dotenv==0.15.0
python-slugify==4.0.1
requests==2.25.1
six==1.15.0
text-unidecode==1.3
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3